+++ deploy.sh
@@ -6,20 +6,5 @@ set -e
 # build
 npm run build
 
-# navigate into the build output directory
-cd dist
-
-# if you are deploying to a custom domain
-# echo 'www.example.com' > CNAME
-
-git init
-git add -A
-git commit -m 'deploy'
-
-# if you are deploying to https://<USERNAME>.github.io
-# git push -f git@github.com:<USERNAME>/<USERNAME>.github.io.git master
-
-# if you are deploying to https://<USERNAME>.github.io/<REPO>
-git push -f git@github.com:andrewtougas/21mafia.git master:gh-pages
-
-cd -
+# copy the build output directory to live server
+rsync -avP ./dist/ atougas@50.116.7.202:/var/www/html/21mafia.andrewtougas.com/public_html/