+++ vue.config.js
@@ -1,5 +1,6 @@
 module.exports = {
-  publicPath: process.env.NODE_ENV === 'production'
-    ? '/21mafia/'
-    : '/'
+  publicPath: process.env.VUE_APP_PUBLIC_PATH
+  // publicPath: process.env.NODE_ENV === 'production'
+  //   ? '/21mafia/'
+  //   : '/'
 }